logify-backend/requirements-dev.txt
-r requirements.txt
black
isort
flake8
pytest
pytest-django
djangorestframework
testcontainers[postgresql]
psycopg2-binary
django-cors-headers
mailjet-rest
